connected client list to store. working on being able to add client to database
