+++ tieme-ndo/src/reducer/index.js
@@ -1,11 +1,16 @@
-
+import { FETCH_DATA } from '../actions/index';
 
 const initialState = {
-    farmers: []
+    farmers: ["No farmer data"]
 }
 
 const reducer = (state = initialState, action) => {
     switch(action.type) {
+        case FETCH_DATA:
+            return {
+                farmers: action.payload
+            }
+
         default:
             return state;
     }
